backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
Werkzeug==3.1.2
